fix: subtle bug causing editor instances to be entangled
The "head" cursor was accidentally globally defined and multiple buffer instances
were referencing the same head. This might explain why the editor would
jump occasionally. One of the symptoms is the first editor seemed to be initialized
as isSelection. I checked no calls to makeSelection was made. Additionally, at
the end of the first instantiation, isSelection was correctly false. It seems
something happened after instantiation, nearly instaneously. Well,
the second instantiation of WarrenBuffer messed with the state of the first one.